domain: retire the sentinel ints — absence is a type, not a value
Same principle as the typed cursors, one layer down. A field whose
comment has to spell out which integers are legal is a type that was
never written down, and every reader re-implements the decoding.

  Thread::rag_mode_override  was `int = -1` documented as "-1 inherit,
      0 On, 1 FirstTurnOnly, 2 Off". Consumers cast the enum back to int
      to compare, and the chain of comparisons ended in an else that
      swallowed every unrecognised value as "inherit" — so a corrupt or
      future integer silently became a policy. Now
      optional<RagMode>: presence and meaning are both in the type, and
      the consumer is a total switch that a new mode will break loudly.

  Model::smart_assign_slot   was `int = -1` with roles as 0/1/2, cast to
      ModelRole on the way out and back to int on the way in — two casts
      around a value that is a role the entire time. Now
      optional<ModelRole>, and the casts are gone.

Storage is unchanged: the key is still absent-or-integer on disk. The
loader now range-checks before converting rather than casting blindly.

This exposed a real layering fault. RagMode lived in store/store.hpp,
but store/ includes domain/conversation.hpp, so a Thread could not name
its own override's type without a cycle. A mode is a fact about a
conversation, not about a file format — it moves to domain/rag_mode.hpp,
with store:: keeping an alias so existing spellings compile. The
alternative (a second identical enum) is precisely the duplication this
work exists to remove.

Removing the sentinels also deleted three more copies of knowledge that
should live in one place: the 0/1/2 switch in the slot-assign writer (now
RoleConfig::slot), the `slot == 0 ? "Strategic" : ...` in the picker
title (now role_display_name — the fourth copy of those labels), and a
`+ 1` in visual_hash that existed only to lift -1 out of negative range.

The compiler found every one of these; that is the point. Verified by
compiling each historical shape against the new types: -1, an
out-of-range integer, and a bare 0/1/2 role are all compile errors now,
while the legitimate uses build.

475/475 green.

+++ src/runtime/app/update/fork.cpp
@@ -139,7 +139,7 @@ Step fork_update(Model m, msg::ForkMsg fm) {
             Thread fork;
             fork.id = deps().new_thread_id();
             fork.forked_from = parent_id;
-            fork.rag_mode_override = static_cast<int>(rag_mode_of(choice));
+            fork.rag_mode_override = rag_mode_of(choice);
             fork.created_at = fork.updated_at = std::chrono::system_clock::now();
             fork.title = m.d.current.title.rfind("Fork: ", 0) == 0
                              ? m.d.current.title
